Move random quote state to main component
